refactor: change components directory structure [MINFOLINK-6]
1. style folder -> styles.tsx
